+++ bot/config/service_locator.py
@@ -1,17 +1,9 @@
-import logging
-
-logger = logging.getLogger(__name__)
-
-
 class ServiceLocator:
     __dependencies = {}
 
     @classmethod
     def register_dependency(cls, dependency_class, service):
         if service in cls.__dependencies:
-            logger.error(
-                f"register_dependency - Dependency {dependency_class.__name__} is already registered"
-            )
             raise Exception(
                 f"Dependency {dependency_class.__name__} is already registered."
             )
@@ -20,13 +12,9 @@ class ServiceLocator:
     @classmethod
     def get_dependency(cls, dependency_class):
         if dependency_class not in cls.__dependencies:
-            logger.error(
-                f"get_dependency - Dependency {dependency_class.__name__} not registered"
-            )
             raise Exception(f"Dependency {dependency_class.__name__} not registered.")
         return cls.__dependencies[dependency_class]
 
     @classmethod
     def clear(cls):
         cls.__dependencies.clear()
-        logger.info("clear - Dependencies cleared")
