+++ bot/domain/utility.py
@@ -1,4 +1,4 @@
-from bot.cog.exception.missing_arguments_exception import MissingArgumentsException
+from bot.resource.exception.missing_arguments_exception import MissingArgumentsException
 
 
 class Utility:
