+++ bot/resource/chain_of_responsibility/handlers/strip_handler.py
@@ -1,4 +1,6 @@
-from bot.cog.chain_of_responsibility.responsibility_handler import ResponsibilityHandler
+from bot.resource.chain_of_responsibility.responsibility_handler import (
+    ResponsibilityHandler,
+)
 
 
 class StripHandler(ResponsibilityHandler):
