+++ bot/resource/chain_of_responsibility/responsibility.py
