+++ bot/resource/chain_of_responsibility/responsibility_builder.py
@@ -1,4 +1,6 @@
-from bot.cog.chain_of_responsibility.responsibility_handler import ResponsibilityHandler
+from bot.resource.chain_of_responsibility.responsibility_handler import (
+    ResponsibilityHandler,
+)
 
 
 class ResponsibilityBuilder:
