+++ bot/resource/cog/association/association.py
@@ -2,9 +2,9 @@ from discord import Message
 from discord.ext import commands
 from discord.ext.commands import Context
 
-from bot.cog.association.embed.aeglo_embed import AegloEmbed
-from bot.cog.association.embed.asetin_embed import AsetinEmbed
-from bot.cog.constants import ReplyMessage
+from bot.resource.cog.association.embed.aeglo_embed import AegloEmbed
+from bot.resource.cog.association.embed.asetin_embed import AsetinEmbed
+from bot.resource.constants import ReplyMessage
 from bot.config.logger.logger import Logger
 from bot.config.service_locator import ServiceLocator
 from bot.domain.discord_client.discord_client import DiscordClient
