+++ bot/resource/cog/association/embed/aeglo_embed.py
