+++ bot/resource/cog/association/embed/asetin_embed.py
