+++ bot/resource/cog/registration/factory/add_student_request_factory.py
@@ -1,6 +1,8 @@
-from bot.cog.chain_of_responsibility.responsibility_handler import ResponsibilityHandler
-from bot.cog.registration.factory.request_factory import RequestFactory
-from bot.cog.registration.request.add_student_request import AddStudentRequest
+from bot.resource.chain_of_responsibility.responsibility_handler import (
+    ResponsibilityHandler,
+)
+from bot.resource.cog.registration.factory.request_factory import RequestFactory
+from bot.resource.cog.registration.request.add_student_request import AddStudentRequest
 
 
 class AddStudentRequestFactory(RequestFactory):
