+++ bot/resource/cog/registration/factory/register_student_request_factory.py
@@ -1,6 +1,10 @@
-from bot.cog.chain_of_responsibility.responsibility_handler import ResponsibilityHandler
-from bot.cog.registration.factory.request_factory import RequestFactory
-from bot.cog.registration.request.register_student_request import RegisterStudentRequest
+from bot.resource.chain_of_responsibility.responsibility_handler import (
+    ResponsibilityHandler,
+)
+from bot.resource.cog.registration.factory.request_factory import RequestFactory
+from bot.resource.cog.registration.request.register_student_request import (
+    RegisterStudentRequest,
+)
 
 
 class RegisterStudentRequestFactory(RequestFactory):
