+++ bot/resource/cog/registration/request/add_student_request.py
