+++ bot/resource/cog/registration/request/register_student_request.py
