+++ bot/resource/constants.py
