+++ bot/resource/decorator/role_check.py
