+++ tools/migrate_discord_students.py
@@ -10,7 +10,7 @@ from discord import Member
 from pymongo import MongoClient
 from pymongo.collection import Collection
 
-from bot.config.dotenv_configuration import DotEnvConfiguration
+from bot.config.context.dotenv_configuration import DotEnvConfiguration
 from bot.domain.discord_client.discord_client import DiscordClient
 from bot.domain.student.attribut.discord_user_id import DiscordUserId
 from bot.domain.student.student import Student
